Revert "Installed Tailwind CSS and TypeScript"
