+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-app",
+  "name": "harmonyhues",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -39,8 +39,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "tailwindcss": "^3.4.7"
   }
 }
