--- tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-      "outDir": "build",
-      "sourceMap": true,
-      "noImplicitAny": true,
-      "moduleResolution": "node",
-      "esModuleInterop": true,
-      "allowSyntheticDefaultImports": true,
-      "jsx": "react-jsx",
-      "target": "esnext",
-      "lib": ["dom", "esnext"],
-      "strict": true
-    },
-    "include": ["src/**/*"]
-}
